fix: method signature

+++ Technical-Department/Technical-Department.Kitchen.API/Public/IIngredientService.cs
@@ -8,6 +8,6 @@ namespace Technical_Department.Kitchen.API.Public
         Result<IngredientDto> GetAllBy();
         Result<IngredientDto> Create(IngredientDto ingredient);
         Result<IngredientDto> Update(IngredientDto ingredient);
-        Result<IngredientDto> Get(long id);
+        Result<IngredientDto> Get(int id);
     }
 }
